infusion/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==23.1.0
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cmake==3.26.4
confection==0.0.4
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
filelock==3.9.0
fire==0.5.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
huggingface-hub==0.15.1
hydra-core==1.3.2
idna==3.4
intertext-graph @ git+https://github.com/UKPLab/intertext-graph.git@2516bd20a7153e825b89206dac0ebeb7d1ab7302
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
lightning-utilities==0.3.0
lit==16.0.6
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.2
mpmath==1.2.1
multidict==6.0.4
murmurhash==1.0.9
networkx==3.0
nltk==3.8.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.1
pandas==2.0.2
pathy==0.10.1
Pillow==9.3.0
preshed==3.0.8
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rouge-score==0.1.2
rsa==4.9
safetensors==0.3.1
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
thinc==8.1.10
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typer==0.7.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wasabi==1.1.2
Werkzeug==2.3.6
yarl==1.9.2
